requirements.txt
aniso8601==9.0.1
blinker==1.8.2
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
passlib==1.7.4
pytz==2024.2
setuptools==75.1.0
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
WTForms==3.1.2
